adding confirm alert before deletion

+++ src/app/employee-list/employee-list.component.ts
@@ -39,10 +39,13 @@ export class EmployeeListComponent implements OnInit {
   }
 
   deleteEmployee(id: number){
-    this.employeeService.deleteEmployee(id).subscribe( data => {
-      console.log(data);
-      this.getEmployees();
-    })
+    let dl = confirm("Are you sure that you want to delete this Question?");
+    if(dl){
+      this.employeeService.deleteEmployee(id).subscribe( data => {
+        console.log(data);
+        this.getEmployees();
+      })
+    }
   }
 
   onSubmitMethod(){
