+++ src/app/javaintro/javaintro.component.ts
@@ -36,10 +36,13 @@ export class JavaintroComponent implements OnInit {
   }
 
   deleteEmployee(id: number){
-    this.employeeService.deleteEmployee(id).subscribe( data => {
-      console.log(data);
-      this.getEmployees();
-    })
+    let dl = confirm("Are you sure that you want to delete this Question?");
+    if(dl){
+      this.employeeService.deleteEmployee(id).subscribe( data => {
+        console.log(data);
+        this.getEmployees();
+      })
+    }
   }
 
 }
